% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestimate.R
\name{anccall}
\alias{anccall}
\title{anccall}
\usage{
anccall(delta = 0.2, window_size = 50000, inputdir, outputdir)
}
\arguments{
\item{delta}{The cutoff for ancestral allele frequency difference}

\item{window_size}{The size of the sliding window}

\item{inputdir}{The file to the ancestral likelihood file}

\item{outputdir}{The file to store the ancestry call}
}
\value{
nothing
}
\description{
anccall
}
