% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anccall_c}
\alias{anccall_c}
\title{anccall_c: Generating ancestry call along the chromosome}
\usage{
anccall_c(
  deltaf,
  window,
  SMAX,
  anclikdir,
  output,
  chrom,
  indiv,
  mode,
  n,
  round1,
  round2,
  round3,
  zero_value
)
}
\arguments{
\item{deltaf}{The cut-off for minimum delta f between two reference populations}

\item{window}{Size of the sliding window}

\item{SMAX}{The number of all snp sites}

\item{anclikdir}{the dir to anclik file}

\item{output}{the dir of the output ancfreq file}
}
\description{
anccall_c: Generating ancestry call along the chromosome
}
