% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anccall_c_nozero_int}
\alias{anccall_c_nozero_int}
\title{anccall_c_nozero_int: Generating ancestry call along the chromosome}
\usage{
anccall_c_nozero_int(
  deltaf,
  window,
  SMAX,
  anclikdir,
  int1,
  int2,
  output,
  chrom,
  indiv,
  mode,
  n
)
}
\arguments{
\item{deltaf}{The cut-off for minimum delta f between two reference populations}

\item{window}{Size of the sliding window}

\item{SMAX}{The number of all snp sites}

\item{anclikdir}{the dir to anclik file}

\item{int1}{output intermediate file 1 call}

\item{int2}{output intermediate file 2 after 1st smoothing}

\item{output}{the dir of the output anccall file}
}
\description{
anccall_c_nozero_int: Generating ancestry call along the chromosome
}
