% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancfreq_c}
\alias{ancfreq_c}
\title{ancfreq_c}
\usage{
ancfreq_c(n, type, pop1, n_pop1, n_pop2, pop2, input, output)
}
\arguments{
\item{n}{an integer, the number of individuals contain in the input}

\item{type}{1 or 2, specifying diploid/haploid data.}

\item{pop1}{a numeric vector of index for reference population 1, sep = ' '}

\item{n_pop1}{number of individual for ref pop 1}

\item{n_pop2}{number of individual for ref pop 2}

\item{pop2}{a numeric vector of index for reference population 2, sep = ' '}

\item{input}{the dir of the input genolik file}

\item{output}{the dir of the output ancfreq file}
}
\description{
Calculating the allele frequency in two different ancestral population
}
