% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancfreq_merge}
\alias{ancfreq_merge}
\title{ancfreq_merge: Merging haploid & dipoid dataset}
\usage{
ancfreq_merge(hapfreq, dipfreq, outputdir, type)
}
\arguments{
\item{hapfreq}{The dir to file containing ancfreq for haploid data}

\item{dipfreq}{The dir to file containing ancfreq for diploid data}

\item{outputdir}{The output dir}

\item{type}{An integer, 0-intersection 1-keep all haploid snp 2-keep all diploid snp 3-union}
}
\description{
ancfreq_merge: Merging haploid & dipoid dataset
}
