% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancgen.R
\name{ancgen}
\alias{ancgen}
\title{ancgen}
\usage{
ancgen(
  genodir,
  ancfreqdir,
  outputdir,
  type = "dip",
  test = "all",
  delta,
  window,
  chrlength,
  mode_n = 0.5,
  min_n = 20,
  exclude = 50000
)
}
\arguments{
\item{genodir}{The directory to genolik file.}

\item{ancfreqdir}{The directory to ancestral frequency file.}

\item{outputdir}{The folder to save output files.}

\item{type}{'dip' for diploid data (default) and 'hap' for haploid data.}

\item{test}{The individuals to be tested. 'all' for all individuals (default), or a numeric vector containing the individual number to be estimated}

\item{delta}{The cutoff for ancestral allele frequency difference}

\item{window}{The size of the sliding window for round 1 and 2, and half of the sliding window for round 3.}

\item{chrlength}{A list of the length to the chromosome.}

\item{mode_n}{The minimum percentage for the mode to make a call, default set to 0.5.}

\item{min_n}{The minimum number of SNPs within the window to make a call, default set to zero.}

\item{exclude}{The length of the start/end of each chromosome to ignore.}
}
\value{
something
}
\description{
A general wrapped-up function for steps after ancestral frequency estimation.
}
