% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anclik.R
\name{anclik}
\alias{anclik}
\title{anclik}
\usage{
anclik(genodir, ancfreqdir, outputdir, type = "dip", test = "all")
}
\arguments{
\item{genodir}{genotype likelihood file}

\item{ancfreqdir}{ancestral allele frequency file}

\item{outputdir}{ancestral state estimation}

\item{type}{'hap' or 'dip'}

\item{test}{can be 'all' or a numeric vector containing the numeric index (e.g. c(1,2,3) will represent the first, second and third individuals in the vcf file) of individuals to be estimated. A suffix "_n" will be added to each output file.}
}
\value{
nothing
}
\description{
anclik
}
