% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anclik_c}
\alias{anclik_c}
\title{anclik_c: Calculating the ancestral likelihood of a test individual}
\usage{
anclik_c(sum, type, testid, genolik, ancfreq, output)
}
\arguments{
\item{sum}{an integer, the number of individuals contain in the input}

\item{type}{1 or 2, specifying diploid/haploid data.}

\item{testid}{the individual to be estimate}

\item{genolik}{the dir to genolik file}

\item{ancfreq}{the dir to ancfreq file}

\item{output}{the dir of the output ancfreq file}
}
\description{
anclik_c: Calculating the ancestral likelihood of a test individual
}
