% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anctract.R
\name{anctract}
\alias{anctract}
\title{anctract}
\usage{
anctract(inputdir, outputdir, chrlength, exclude = 0)
}
\arguments{
\item{inputdir}{input file directory}

\item{outputdir}{output file directory}

\item{chrlength}{length of the chrmosome}

\item{exclude}{length to be excluded from both end}
}
\description{
Generating tract from call
}
