% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anctract_c}
\alias{anctract_c}
\title{anctract_c: Generating ancestry tract from call.}
\usage{
anctract_c(chrlen, excludelen, input, output)
}
\arguments{
\item{chrlen}{Length of the chromosome}

\item{excludelen}{length to be excluded}

\item{input}{input file dir}

\item{output}{output file dir}
}
\description{
anctract_c: Generating ancestry tract from call.
}
