% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck.R
\name{datacheck}
\alias{datacheck}
\title{datacheck
A function to check if the file contains certain wrong character types}
\usage{
datacheck(input, character = "[^0-9[:space:]]", field = "1-")
}
\arguments{
\item{input}{the input file dir}

\item{character}{the character types regular expression i.e. '[^0-9]'}

\item{field}{the field to be kept, default for all}
}
\value{
0 pass 1 error

datacheck(genolik,'[^0-9[:space:]]','2-')
}
\description{
datacheck
A function to check if the file contains certain wrong character types
}
