% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck.R
\name{freqsum}
\alias{freqsum}
\title{freqsum}
\usage{
freqsum(freq)
}
\arguments{
\item{freq}{The dir to ancestral allele frequency file}
}
\value{
A numeric vector of the difference of allele frequency
}
\description{
Basic summary for ancestral frequency file
}
