% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtbaboon1b.R
\name{get_chrlist}
\alias{get_chrlist}
\title{get_chrlist}
\usage{
get_chrlist(
  vcf = "/Users/lycium/Desktop/Jennylab/rpackage_LCLAE/rawtestdata/test_clean.vcf"
)
}
\arguments{
\item{vcf}{the input vcf file}
}
\value{
a character vector rep,chr
}
\description{
get_chrlist
}
