% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtbaboon1b.R
\name{get_genolik}
\alias{get_genolik}
\title{get_genolik: vcf to genolik}
\usage{
get_genolik(input, output, type = "PL")
}
\arguments{
\item{input}{vcf}

\item{output}{genolik}
}
\value{
nothing
}
\description{
get_genolik: vcf to genolik
}
