% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_genolik_c}
\alias{get_genolik_c}
\title{get_genolik_c: Get genolik directly from vcf}
\usage{
get_genolik_c(input_file, output_file)
}
\arguments{
\item{input_file}{a string, the dir to cleaned vcf file}

\item{output_file}{a string, the dir to the output file}
}
\description{
get_genolik_c: Get genolik directly from vcf
}
