% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calltotract.R
\name{gettract}
\alias{gettract}
\title{gettract}
\usage{
gettract(
  datadir,
  outputdir,
  indiv_name = "50",
  chr = "chrX",
  chrlength = "142711496",
  value = 35000,
  mode_n = 0.5,
  min_n = 20,
  exclude = 50000
)
}
\arguments{
\item{datadir}{input snp level ancestry call file}

\item{outputdir}{output tract file}

\item{indiv_name}{The unique identifier for the individual, which could be numeric index (e.g. 1) or character sample name (e.g. "sample_1")}

\item{chr}{the chromosome it is on}

\item{chrlength}{the length of the chromosome}

\item{value}{window size}

\item{mode_n}{the minimum percentage for a call to be made using majority rule}

\item{min_n}{the minimum number of SNPs within the window to make a call}

\item{exclude}{how much of the start/end of each chromosome to ignore}
}
\value{
An output file including tract and majority rule
}
\description{
Generating tract from snp level ancestry call.
}
