% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{inputcheck}
\alias{inputcheck}
\title{inputcheck}
\usage{
inputcheck(inputdir)
}
\arguments{
\item{inputdir}{dir to clean vcf}
}
\value{
A string indicating the result of the input check
}
\description{
A function to check the format of the input cleaned-up vcf file
}
