% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancfreq.R
\name{mergefreq}
\alias{mergefreq}
\title{mergefreq}
\usage{
mergefreq(hap, dip, merge, type = "union")
}
\arguments{
\item{hap}{The ancestral frequency calculated for haploid data}

\item{dip}{The ancestral frequency calculated for diploid data}

\item{merge}{The output directory for merged ancestral frequency}

\item{type}{A parameter for merging type ("intersect","union","full_dip","full_hap"). 'union' for default.}
}
\value{
A merged ancfreq file
}
\description{
A function to merge existing ancestral frequency files
}
