% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(inputdir, outputdir)
}
\arguments{
\item{inputdir}{dir to orginial vcf file}

\item{outputdir}{dir to store cleaned_up vcf file}
}
\description{
A function to preprocess the vcf file
}
