% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLCLAE_pip.R
\name{rLCLAE}
\alias{rLCLAE}
\title{rLCLAE}
\usage{
rLCLAE(
  input,
  tmp,
  output,
  chrlist,
  refpop1_dip,
  refpop2_dip,
  test_ind = "all",
  inputtype = "dip",
  freqtype = "dip",
  window = 50000,
  mode_min = 0.3,
  n_min = 10,
  delta = 0.1,
  exclude = 1000,
  inputdir_hap = "missing",
  refpop1_hap = "missing",
  refpop2_hap = "missing",
  mergetype = "union",
  zero_value = 10^(-7)
)
}
\arguments{
\item{input}{vcf input}

\item{tmp}{intermediate file storage}

\item{output}{final output (not yet used)}

\item{chrlist}{a data frame for chr name & length, from fai}

\item{refpop1_dip}{numeric vector for index of ref ind for pop1, need expand}

\item{refpop2_dip}{numeric vector for index of ref ind for pop2, need expand}

\item{test_ind}{numeric vector for index of test individual, default all}

\item{inputtype}{hap or dip}

\item{freqtype}{hap, dip or merge}

\item{window}{window size}

\item{mode_min}{mode min}

\item{n_min}{n min in window}

\item{delta}{freq diff lower cutoff}

\item{exclude}{exclude length}

\item{inputdir_hap}{missing, do not touch, not finished for ancfreq}

\item{refpop1_hap}{missing, do not touch, not finished for ancfreq}

\item{refpop2_hap}{missing, do not touch, not finished for ancfreq}

\item{mergetype}{union, do not touch, not finished for ancfreq}

\item{zero_value}{a small zero value added to the zeros, if zero then throw out sites with any zero}
}
\description{
The summarized pipeline for rLCLAE
}
