% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestimate.R
\name{anccall}
\alias{anccall}
\title{anccall}
\usage{
anccall(
  inputdir,
  outputdir,
  chrom_name,
  indiv_name,
  mode_min = 0.5,
  n_min = 20,
  delta = 0.2,
  window_size = 50000,
  round1 = 1,
  round2 = 1,
  round3 = 2,
  auto_optimize = F,
  ploidy = "dip",
  chrlength = 142711496,
  zero_value = 10^(-7)
)
}
\arguments{
\item{inputdir}{The file to the ancestral likelihood file}

\item{outputdir}{The file to store the ancestry call}

\item{chrom_name}{The name of the chromosome}

\item{indiv_name}{The name of the individual}

\item{mode_min}{The minimum percentage for the mode to make a call, default set to 0.5}

\item{n_min}{the minimum number of SNPs within the window to make a call, default set to zero}

\item{delta}{The cutoff for ancestral allele frequency difference}

\item{window_size}{The size of the sliding window}

\item{round3}{The scaler for window_size for third round}

\item{ploidy}{"hap" or "dip", default set to "dip"}

\item{chrlength}{chromosome length, default set to baboon X: 142711496}

\item{zero_value}{The small value added to avoid zeroes, default to 10^-7}
}
\value{
nothing
}
\description{
anccall
}
