% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancfreq.R
\name{ancfreq}
\alias{ancfreq}
\title{ancfreq}
\usage{
ancfreq(
  outputdir,
  inputdir_dip = "missing",
  pop1_dip = "missing",
  pop2_dip = "missing",
  inputdir_hap = "missing",
  pop1_hap = "missing",
  pop2_hap = "missing",
  mergetype = "union"
)
}
\arguments{
\item{outputdir}{The dir to the output file. A postfix ('_dip','_hap','_merge') will be added.}

\item{inputdir_dip}{The dir to a file containing diploid genotype likelihood data.}

\item{pop1_dip}{A numeric vector specifying the index of reference population 1 for diploid data.}

\item{pop2_dip}{A numeric vector specifying the index of reference population 2 for diploid data.}

\item{inputdir_hap}{The dir to a file containing haploid genotype likelihood data.}

\item{pop1_hap}{A numeric vector specifying the index of reference population 1 for haploid data.}

\item{pop2_hap}{A numeric vector specifying the index of reference population 2 for haploid data.}

\item{mergetype}{A parameter for merging type ("intersect","union","full_dip","full_hap")}
}
\value{
The output for ancfreq consists of a single line for each SNP. The first field shows the base position; the second to third fields show the estimated allele frequency for reference population 1 and 2; the third field shows the estimated difference in allele frequency in the two reference populations, the fourth to fifth fields show available individuals in reference population one and two.
}
\description{
A function to calculate ancestral allele frequency based on two reference populations.
}
