% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancfreq_c}
\alias{ancfreq_c}
\title{ancfreq_c: Calculating the allele frequency in two different ancestral population}
\usage{
ancfreq_c(n, type, pop1, pop2, input, output)
}
\arguments{
\item{n}{an integer, the number of individuals contain in the input}

\item{type}{1 or 2, specifying diploid/haploid data.}

\item{pop1}{the dir of a file containing individuals belonging to reference population 1}

\item{pop2}{the dir of a file containing individuals belonging to reference population 2}

\item{input}{the dir of the input genolik file}

\item{output}{the dir of the output ancfreq file}
}
\description{
ancfreq_c: Calculating the allele frequency in two different ancestral population
}
