% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anclik.R
\name{anclik}
\alias{anclik}
\title{anclik}
\usage{
anclik(genodir, ancfreqdir, outputdir, type = "dip", test)
}
\arguments{
\item{genodir}{genotype likelihood file}

\item{ancfreqdir}{ancestral allele frequency file}

\item{outputdir}{ancestral state estimation}

\item{type}{'hap' or 'dip'}

\item{test}{the individual number to be estimated}
}
\value{
nothing
}
\description{
anclik
}
