% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtbaboon1b.R
\name{genolik}
\alias{genolik}
\title{Calculating the relative genotype likelihood from a vcf file.}
\usage{
genolik(inputdir, outputdir)
}
\arguments{
\item{inputdir}{The input directory of the cleaned vcf file.}

\item{outputdir}{The output directory of the genolik file.}
}
\value{
A file in which each line represents a snp. The line looks like chr, pos, -1. -1. -1. (missing data) or gl1, gl2, gl3 transformed from the phred score.
}
\description{
Calculating the relative genotype likelihood from a vcf file.
}
