% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{inputcheck}
\alias{inputcheck}
\title{inputcheck
A function to check the format of the input cleaned-up vcf file}
\usage{
inputcheck(inputdir)
}
\arguments{
\item{inputdir}{dir to clean vcf}
}
\value{
nothing
}
\description{
inputcheck
A function to check the format of the input cleaned-up vcf file
}
